Added pose estimate param defaults. (#13)
The initial goal of this branch was to ensure that the estimator outputs the pose of the marker relative to the camera. The [ArUco tutorial](https://docs.opencv.org/4.x/d5/dae/tutorial_aruco_detection.html) seems to say that the function `cv::aruco::estimatePoseSingleMarkers` estimates the pose of the camera relative to the marker. However, the [documentation for the function](https://docs.opencv.org/4.x/d9/d6a/group__aruco.html#ga56878c4f94362c9420f20b78de0def9b) says that the function estimates the pose of the marker relative to the camera. Personal experimentation verifies that the function estimates the pose of the marker relative to the camera.

To ensure a specific default behavior, I added estimation parameters to `estimatePoseSingleMarkers` function in `ArucoMarkerEstimator`.